requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0,<11.0.0
boto3>=1.26.0
pytest>=6.0.0
python-dotenv>=0.19.0
typing-extensions>=4.0.0
numpy>=1.24.0
pillow>=9.5.0